requirements.txt
# brightway2
arrow
fastapi
openai
pinecone-client
python-dotenv
tenacity
tiktoken
uvicorn[standard]
xata